# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cadit",
        "contact": {
            "name": "Mingeun Kim (Kevin)",
            "url": "https://mingeun.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/users/bank": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "\uc774\uc6a9\uc790\uac00 \uc740\ud589 \uacc4\uc88c\ub97c \ub4f1\ub85d\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uac00 \uc740\ud589 \uacc4\uc88c\ub97c \ub4f1\ub85d\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "bank_name",
                        "in": "query",
                        "description": "User Bank Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "User Birthday",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "bank_account_number",
                        "in": "query",
                        "description": "User Bank Account Number",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "bank_account_password",
                        "in": "query",
                        "description": "User Bank Account Password first and second",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful User Bank Account"
                    }
                }
            }
        },
        "/users/card": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "\uc774\uc6a9\uc790\uac00 \uce74\ub4dc\ub97c \ub4f1\ub85d\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uac00 \uce74\ub4dc\ub97c \ub4f1\ub85d\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "card_no",
                        "in": "query",
                        "description": "User Card Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "card_valid_month",
                        "in": "query",
                        "description": "User Card Valid Month",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "card_valid_year",
                        "in": "query",
                        "description": "User Card Valid Year",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "card_valid_cvc",
                        "in": "query",
                        "description": "User Card Valid CVC",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Card Verify"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\uc774\uc6a9\uc790\ub97c \ucd94\uac00(\ud68c\uc6d0\uac00\uc785) \ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\ub97c \ucd94\uac00(\ud68c\uc6d0\uac00\uc785) \ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User Phone NUmber",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Verify Phone Code",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Register User"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\uc774\uc6a9\uc790\uac00 \ub85c\uadf8\uc778\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uac00 \ub85c\uadf8\uc778\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Login User"
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\uc774\uc6a9\uc790\uc758 \uc0c1\uc138\uc815\ubcf4\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uc758 \uc0c1\uc138\uc815\ubcf4\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful User Information"
                    }
                }
            }
        },
        "/auth/token/refresh": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\uc774\uc6a9\uc790\uc758 \uc778\uc99d \ud1a0\ud070\uc744 \uc7ac\ubc1c\uae09(\uac31\uc2e0) \ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uc758 \uc778\uc99d \ud1a0\ud070\uc744 \uc7ac\ubc1c\uae09(\uac31\uc2e0) \ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful User Refresh JWT Token"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\uc774\uc6a9\uc790\ub97c \ub85c\uadf8\uc544\uc6c3 \ucc98\ub9ac\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\ub97c \ub85c\uadf8\uc544\uc6c3 \ucc98\ub9ac\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful User Logout"
                    }
                }
            }
        },
        "/auth/phone": {
            "post": {
                "tags": [
                    "Verify"
                ],
                "summary": "\uc774\uc6a9\uc790\uac00 \ud68c\uc6d0\uac00\uc785 \uc2dc \ud734\ub300\ud3f0 \uc778\uc99d \ubc88\ud638\ub97c \uc804\uc1a1\ud569\ub2c8\ub2e4.",
                "description": "\uc774\uc6a9\uc790\uac00 \ud68c\uc6d0\uac00\uc785 \uc2dc \ud734\ub300\ud3f0 \uc778\uc99d \ubc88\ud638\ub97c \uc804\uc1a1\ud569\ub2c8\ub2e4.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User Phone Number",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful Phone Verify Send"
                    }
                }
            }
        }
    },
    "definitions": {}
}